# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Budgie",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/expenses": {
      "get": {
        "tags": [
          "expenses"
        ],
        "summary": "List expenses, newest date first",
        "operationId": "list_expenses_api_expenses_get",
        "parameters": [
          {
            "name": "review_status",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "enum": [
                    "accepted",
                    "needs_review",
                    "rejected"
                  ],
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Review Status"
            }
          },
          {
            "name": "confirmed",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "boolean"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Confirmed"
            }
          },
          {
            "name": "from",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "date"
                },
                {
                  "type": "null"
                }
              ],
              "description": "inclusive",
              "title": "From"
            },
            "description": "inclusive"
          },
          {
            "name": "to",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "date"
                },
                {
                  "type": "null"
                }
              ],
              "description": "inclusive",
              "title": "To"
            },
            "description": "inclusive"
          },
          {
            "name": "category",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "enum": [
                    "groceries.fresh",
                    "groceries.staples",
                    "snacks_sweets",
                    "drinks",
                    "alcohol",
                    "tobacco",
                    "household",
                    "personal_care",
                    "health",
                    "eating_out",
                    "transport",
                    "clothing",
                    "electronics",
                    "other",
                    "deposit",
                    "discount",
                    "uncategorized"
                  ],
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "expenses with at least one item in this category",
              "title": "Category"
            },
            "description": "expenses with at least one item in this category"
          },
          {
            "name": "has_receipt",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "boolean"
                },
                {
                  "type": "null"
                }
              ],
              "description": "false: only expenses without a receipt (manual); true: with one",
              "title": "Has Receipt"
            },
            "description": "false: only expenses without a receipt (manual); true: with one"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Expense"
                  },
                  "title": "Response List Expenses Api Expenses Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation or business-rule error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "500": {
            "description": "Server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "expenses"
        ],
        "summary": "Create a manual expense, optionally for a failed receipt",
        "operationId": "create_expense_api_expenses_post",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ExpenseCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Expense"
                }
              }
            }
          },
          "422": {
            "description": "Validation or business-rule error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "500": {
            "description": "Server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "404": {
            "description": "Not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "409": {
            "description": "Invalid state",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          }
        }
      }
    },
    "/api/expenses/export.csv": {
      "get": {
        "tags": [
          "expenses"
        ],
        "summary": "Confirmed expenses as CSV, one row per line item",
        "operationId": "export_csv_api_expenses_export_csv_get",
        "parameters": [
          {
            "name": "from",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "date"
                },
                {
                  "type": "null"
                }
              ],
              "description": "inclusive",
              "title": "From"
            },
            "description": "inclusive"
          },
          {
            "name": "to",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "date"
                },
                {
                  "type": "null"
                }
              ],
              "description": "inclusive",
              "title": "To"
            },
            "description": "inclusive"
          }
        ],
        "responses": {
          "200": {
            "description": "CSV file (contracts/csv-export.md)",
            "content": {
              "text/csv": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "422": {
            "description": "Validation or business-rule error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "500": {
            "description": "Server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          }
        }
      }
    },
    "/api/expenses/{id}": {
      "get": {
        "tags": [
          "expenses"
        ],
        "summary": "Get an expense",
        "operationId": "get_expense_api_expenses__id__get",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Expense"
                }
              }
            }
          },
          "422": {
            "description": "Validation or business-rule error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "500": {
            "description": "Server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "404": {
            "description": "Not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": [
          "expenses"
        ],
        "summary": "Edit an expense; a confirmed expense becomes unconfirmed",
        "operationId": "update_expense_api_expenses__id__patch",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ExpenseUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Expense"
                }
              }
            }
          },
          "422": {
            "description": "Validation or business-rule error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "500": {
            "description": "Server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "404": {
            "description": "Not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "expenses"
        ],
        "summary": "Delete an expense, its receipt and its image",
        "operationId": "delete_expense_api_expenses__id__delete",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation or business-rule error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "500": {
            "description": "Server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "404": {
            "description": "Not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          }
        }
      }
    },
    "/api/expenses/{id}/confirm": {
      "post": {
        "tags": [
          "expenses"
        ],
        "summary": "Confirm an expense",
        "operationId": "confirm_expense_api_expenses__id__confirm_post",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Expense"
                }
              }
            }
          },
          "422": {
            "description": "Validation or business-rule error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "500": {
            "description": "Server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "404": {
            "description": "Not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ErrorBody": {
        "properties": {
          "error": {
            "type": "string",
            "enum": [
              "bad_request",
              "not_found",
              "method_not_allowed",
              "invalid_state",
              "file_too_large",
              "unsupported_file",
              "uncategorized_items",
              "incomplete_expense",
              "validation_error",
              "storage_error",
              "internal_error",
              "not_implemented"
            ],
            "title": "Error"
          },
          "detail": {
            "type": "string",
            "title": "Detail"
          },
          "fields": {
            "anyOf": [
              {
                "items": {
                  "$ref": "#/components/schemas/FieldError"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Fields"
          }
        },
        "type": "object",
        "required": [
          "error",
          "detail"
        ],
        "title": "ErrorBody",
        "description": "The body of every non-2xx response (error-format.md)."
      },
      "Expense": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "receipt_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Receipt Id"
          },
          "merchant": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Merchant"
          },
          "date": {
            "anyOf": [
              {
                "type": "string",
                "format": "date"
              },
              {
                "type": "null"
              }
            ],
            "title": "Date"
          },
          "currency": {
            "type": "string",
            "pattern": "^[A-Z]{3}$",
            "title": "Currency",
            "examples": [
              "EUR"
            ]
          },
          "total": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Total"
          },
          "subtotal": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Subtotal"
          },
          "tax": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Tax"
          },
          "source": {
            "type": "string",
            "enum": [
              "ai",
              "ai_corrected",
              "manual"
            ],
            "title": "Source"
          },
          "review_status": {
            "type": "string",
            "enum": [
              "accepted",
              "needs_review",
              "rejected"
            ],
            "title": "Review Status"
          },
          "confirmed": {
            "type": "boolean",
            "title": "Confirmed"
          },
          "flags": {
            "items": {
              "$ref": "#/components/schemas/Flag"
            },
            "type": "array",
            "title": "Flags"
          },
          "line_items": {
            "items": {
              "$ref": "#/components/schemas/LineItem"
            },
            "type": "array",
            "title": "Line Items"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "receipt_id",
          "merchant",
          "date",
          "currency",
          "total",
          "subtotal",
          "tax",
          "source",
          "review_status",
          "confirmed",
          "flags",
          "line_items",
          "created_at"
        ],
        "title": "Expense"
      },
      "ExpenseCreate": {
        "properties": {
          "receipt_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Receipt Id"
          },
          "merchant": {
            "type": "string",
            "title": "Merchant"
          },
          "date": {
            "type": "string",
            "format": "date",
            "title": "Date"
          },
          "currency": {
            "type": "string",
            "pattern": "^[A-Z]{3}$",
            "title": "Currency",
            "default": "EUR",
            "examples": [
              "EUR"
            ]
          },
          "total": {
            "type": "number",
            "title": "Total"
          },
          "subtotal": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Subtotal"
          },
          "tax": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Tax"
          },
          "line_items": {
            "items": {
              "$ref": "#/components/schemas/LineItemInput"
            },
            "type": "array",
            "minItems": 1,
            "title": "Line Items"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "merchant",
          "date",
          "total",
          "line_items"
        ],
        "title": "ExpenseCreate",
        "description": "`POST /expenses`: a manual expense, optionally attached to a failed receipt."
      },
      "ExpenseUpdate": {
        "properties": {
          "merchant": {
            "type": "string",
            "title": "Merchant"
          },
          "date": {
            "type": "string",
            "format": "date",
            "title": "Date"
          },
          "currency": {
            "type": "string",
            "pattern": "^[A-Z]{3}$",
            "title": "Currency",
            "examples": [
              "EUR"
            ]
          },
          "total": {
            "type": "number",
            "title": "Total"
          },
          "subtotal": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Subtotal"
          },
          "tax": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Tax"
          },
          "line_items": {
            "items": {
              "$ref": "#/components/schemas/LineItemInput"
            },
            "type": "array",
            "minItems": 1,
            "title": "Line Items"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "title": "ExpenseUpdate",
        "description": "`PATCH /expenses/{id}`: only the fields sent change.\n\nFields that are required in `ExpenseCreate` may be left out but not sent as `null`.\n`line_items`, if sent, replaces the list."
      },
      "FieldError": {
        "properties": {
          "field": {
            "type": "string",
            "title": "Field",
            "examples": [
              "line_items.0.amount"
            ]
          },
          "message": {
            "type": "string",
            "title": "Message"
          }
        },
        "type": "object",
        "required": [
          "field",
          "message"
        ],
        "title": "FieldError",
        "description": "One failed field of a `validation_error`."
      },
      "Flag": {
        "properties": {
          "field": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Field"
          },
          "code": {
            "type": "string",
            "title": "Code"
          },
          "message": {
            "type": "string",
            "title": "Message"
          }
        },
        "type": "object",
        "required": [
          "field",
          "code",
          "message"
        ],
        "title": "Flag",
        "description": "A rule result on an expense, e.g. `sum_mismatch` on `total`."
      },
...
      "LineItem": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "description": {
            "type": "string",
            "title": "Description"
          },
          "normalized_name": {
            "type": "string",
            "title": "Normalized Name"
          },
          "qty": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Qty"
          },
          "unit": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Unit"
          },
          "unit_price": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Unit Price"
          },
          "amount": {
            "type": "number",
            "title": "Amount"
          },
          "category": {
            "type": "string",
            "enum": [
              "groceries.fresh",
              "groceries.staples",
              "snacks_sweets",
              "drinks",
              "alcohol",
              "tobacco",
              "household",
              "personal_care",
              "health",
              "eating_out",
              "transport",
              "clothing",
              "electronics",
              "other",
              "deposit",
              "discount",
              "uncategorized"
            ],
            "title": "Category"
          },
          "category_source": {
            "type": "string",
            "enum": [
              "seed",
              "user",
              "none"
            ],
            "title": "Category Source"
          }
        },
        "type": "object",
        "required": [
          "id",
          "description",
          "normalized_name",
          "qty",
          "unit",
          "unit_price",
          "amount",
          "category",
          "category_source"
        ],
        "title": "LineItem"
      },
      "LineItemInput": {
        "properties": {
          "id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Id"
          },
          "description": {
            "type": "string",
            "title": "Description"
          },
          "qty": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Qty"
          },
          "unit": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Unit"
          },
          "unit_price": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Unit Price"
          },
          "amount": {
            "type": "number",
            "title": "Amount"
          },
          "category": {
            "anyOf": [
              {
                "type": "string",
                "enum": [
                  "groceries.fresh",
                  "groceries.staples",
                  "snacks_sweets",
                  "drinks",
                  "alcohol",
                  "tobacco",
                  "household",
                  "personal_care",
                  "health",
                  "eating_out",
                  "transport",
                  "clothing",
                  "electronics",
                  "other",
                  "deposit",
                  "discount"
                ]
              },
              {
                "type": "null"
              }
            ],
            "title": "Category"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "description",
          "amount"
        ],
        "title": "LineItemInput",
        "description": "A line item as sent by the client; the server derives `normalized_name`."
      },
...
    }
  }
}